Move autostart-functions to new file. Fewer f-strings.
