+++ src/vorta/utils.py
@@ -11,7 +11,6 @@ from collections import defaultdict
 from functools import reduce
 import operator
 import psutil
-from pathlib import Path
 
 from paramiko.rsakey import RSAKey
 from paramiko.ecdsakey import ECDSAKey
@@ -20,8 +19,6 @@ from paramiko import SSHException
 from PyQt5.QtWidgets import QFileDialog
 from PyQt5.QtGui import QIcon
 from PyQt5 import QtCore
-from setuptools import Distribution
-from setuptools.command.install import install
 import subprocess
 import keyring
 from vorta.keyring_db import VortaDBKeyring
@@ -225,50 +222,6 @@ def uses_dark_mode():
     return None
 
 
-def open_app_at_startup(enabled=True):
-    """
-    This function adds/removes the current app bundle from Login items in macOS
-    """
-    if sys.platform == 'darwin':
-        from Foundation import NSDictionary
-
-        from Cocoa import NSBundle, NSURL
-        from CoreFoundation import kCFAllocatorDefault
-        # CF = CDLL(find_library('CoreFoundation'))
-        from LaunchServices import (LSSharedFileListCreate, kLSSharedFileListSessionLoginItems,
-                                    LSSharedFileListInsertItemURL, kLSSharedFileListItemHidden,
-                                    kLSSharedFileListItemLast, LSSharedFileListItemRemove)
-
-        app_path = NSBundle.mainBundle().bundlePath()
-        url = NSURL.alloc().initFileURLWithPath_(app_path)
-        login_items = LSSharedFileListCreate(kCFAllocatorDefault, kLSSharedFileListSessionLoginItems, None)
-        props = NSDictionary.dictionaryWithObject_forKey_(True, kLSSharedFileListItemHidden)
-
-        new_item = LSSharedFileListInsertItemURL(login_items, kLSSharedFileListItemLast,
-                                                 None, None, url, props, None)
-        if not enabled:
-            LSSharedFileListItemRemove(login_items, new_item)
-    elif sys.platform.startswith('linux'):
-        config_path = QtCore.QStandardPaths.writableLocation(QtCore.QStandardPaths.ConfigLocation)
-        autostart_file_path = Path(config_path) / 'autostart' / 'vorta.desktop'
-        if enabled:
-            dir_entry_point = get_setuptools_script_dir()
-            autostart_file_content = (f"[Desktop Entry]\n"
-                                      f"Name=Vorta\n"
-                                      f"GenericName=Backup Software\n"
-                                      f"Exec={dir_entry_point}/vorta\n"
-                                      f"Terminal=false\n"
-                                      f"Icon=vorta\n"
-                                      f"Categories=Utility\n"
-                                      f"Type=Application\n"
-                                      f"StartupNotify=false\n"
-                                      f"X-GNOME-Autostart-enabled=true\n")
-            autostart_file_path.write_text(autostart_file_content)
-        else:
-            if autostart_file_path.exists():
-                autostart_file_path.unlink()
-
-
 def format_archive_name(profile, archive_name_tpl):
     """
     Generate an archive name. Default:
@@ -309,22 +262,3 @@ def get_mount_points(repo_url):
                     break
 
     return mount_points
-
-
-# Get entry point of vorta
-# From https://stackoverflow.com/questions/
-#      25066084/get-entry-point-script-file-location-in-setuputils-package
-class OnlyGetScriptPath(install):
-    def run(self):
-        # does not call install.run() by design
-        self.distribution.install_scripts = self.install_scripts
-
-
-def get_setuptools_script_dir():
-    dist = Distribution({'cmdclass': {'install': OnlyGetScriptPath}})
-    dist.dry_run = True  # not sure if necessary, but to be safe
-    dist.parse_config_files()
-    command = dist.get_command_obj('install')
-    command.ensure_finalized()
-    command.run()
-    return dist.install_scripts
